Add essential info to CLIENT_COMPAT
It took me ages to find out where to get the ovpn config file for my eduVPN connection, and some more searching for a way to import it into NetworkManager. Given that this is something each individual user has to do, I thought it should be a little more obvious.

Is it truly impossible to download a config file for a different country with just a browser? Perhaps servers would need a different WAYF that lets one choose countries first?

Info for Debian 8 / Ubuntu 16.04 is probably obsolete by now, and I doubt there are a lot of new client installations on Debian 9 / Ubuntu 18.04 even...